+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Used keywords can be: `Added`, `Changed`, `Removed`, `Fixed`
 - fix file build order
 - get week info for sheets
 - do not create events for open slots
+- add code to skip weeks other than the current
 
 ## [0.0.2] - 2023-03-31